refactor(core): split providers, sandbox, skills into directory modules (<1000)
- providers.rs (1896) → providers/: model-list fetching (models.rs), OAuth
  device flow + Copilot session exchange (device_flow.rs), tests (tests.rs);
  mod.rs keeps the ProviderDef table + helpers + ModelInfo.
- sandbox.rs (1341) → sandbox/: platform wrappers (bwrap/seatbelt/landlock)
  in platform.rs, tests in tests.rs; mod.rs keeps capabilities/policy/mode/
  path-validation.
- skills.rs (1810) → skills/: ClawHub registry types + SkillManager registry
  methods in clawhub.rs, tests in tests.rs; mod.rs keeps the core SkillManager.

A few private helpers became pub(crate) so the relocated tests can reach them.
All files under 1000 lines. No behavior change.
